app-infra/requirements.txt
aws_cdk.core
aws_cdk.aws_apigateway
aws_cdk.aws_lambda
aws_cdk.aws_iam
aws_cdk.aws_logs
aws_cdk.aws_dynamodb
aws_cdk.aws_s3
aws_cdk.aws_pinpoint